changing flag file path to /tmp
Since /tmp is emptied on boot, and /var/tmp is not, the old setup
resulted in a nasty side-effect: if you were on wifi, then rebooted,
it would assume it should be on ethernet now, or at least that wifi
should be defaulted to off. So on bootup you'd have no wifi, and
if ethernet wasn't plugged in, you'd also have no ethernet.